add BurgerNav.js
