+++ src/components/App/App.css
@@ -14,4 +14,6 @@
     -webkit-text-size-adjust: 100%;
     -ms-text-size-adjust: 100%;
     -moz-text-size-adjust: 100%;
+
+    overflow: hidden;
 }
