+++ src/components/App/App.js
@@ -2,18 +2,23 @@ import {Route, Switch} from 'react-router-dom';
 
 import './App.css';
 
-import Main from '../Main/Main';
-import Header from '../Header/Header';
-import Movies from '../Movies/Movies';
+import Main from '../Pages/Main/Main';
+import Movies from '../Pages/Movies/Movies';
+import SignIn from '../Pages/Auth/SignIn';
+import Signup from '../Pages/Auth/SignUp';
+import Header from "../Sections/Header/Header";
 
 function App() {
     return (
         <div className="root">
             <Header/>
-
             <Switch>
-                <Route path='signin'>
+                <Route path='/signin'>
+                    <SignIn/>
+                </Route>
 
+                <Route path='/signup'>
+                    <Signup/>
                 </Route>
 
                 <Route exact path='/'>
@@ -21,7 +26,7 @@ function App() {
                 </Route>
 
                 <Route path='/movies'>
-                    <Movies />
+                    <Movies/>
                 </Route>
             </Switch>
         </div>
