--- src/components/Navigation/Navigation.scss
@@ -1,38 +1,0 @@
-.navigation {
-  width: 132px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  @media (min-width: 568px) {
-    width: 183px;
-  }
-
-  &__link {
-    font-weight: 500;
-    font-size: 10px;
-    line-height: 16px;
-
-    @media (min-width: 568px) {
-      font-size: 12px;
-      line-height: 16px;
-    }
-  }
-}
-
-.navigation-main {
-
-  &__menu {
-    width: 40px;
-    height: 43px;
-
-    background-image: url('../../images/icon-main-burger.svg');
-    background-color: transparent;
-
-    border: none;
-
-    cursor: pointer;
-  }
-}
