+++ src/components/Other/Buttons/AuthButton/AuthButton.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './AuthButton.scss';
+
+function AuthButton({text}) {
+    return (
+        <button className='auth-button'>{text}</button>
+    );
+}
+
+export default AuthButton;
