+++ src/components/Other/Buttons/GrayButton/GrayButton.scss
