+++ src/components/Other/Inputs/AuthInput/AuthInput.scss
@@ -1,0 +1,41 @@
+.auth-input {
+  margin-top: 18px;
+  max-width: 100%;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+  &__text {
+    margin: 0 0 5px;
+
+    font-weight: 400;
+    font-size: 10px;
+    line-height: 12px;
+    color: #A0A0A0;
+  }
+
+  &__input {
+    padding: 0 0 10px;
+    margin-bottom: 6px;
+    width: 100%;
+
+    border: none;
+    border-bottom: 1px #e8e8e8 solid;
+
+    font-weight: 500;
+    font-size: 14px;
+    line-height: 17px;
+
+    &:focus {
+      outline: none;
+    }
+  }
+
+  &__span {
+    font-weight: 400;
+    font-size: 10px;
+    line-height: 12px;
+    color: #ff4062;
+  }
+}
