+++ src/components/Other/Inputs/FilterCheckbox/FilterCheckbox.scss
