+++ src/components/Other/Inputs/SearchInput/SearchInput.scss
