+++ src/components/Pages/Auth/Auth.scss
@@ -1,0 +1,91 @@
+.auth {
+  margin-left: 30px;
+  margin-right: 30px;
+  margin-bottom: 30px;
+
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  @media (min-width: 568px) {
+    margin-left: 24%;
+    margin-right: 24%;
+    margin-bottom: 22vh;
+
+    align-items: flex-start;
+  }
+
+  @media (min-width: 1280px) {
+    margin-left: 34%;
+    margin-right: 34%;
+    margin-bottom: 70px;
+  }
+
+  &__header {
+    margin-top: 56px;
+    margin-bottom: 80px;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    @media (min-width: 568px) {
+      margin-top: 22vh;
+      margin-bottom: 40px;
+
+      align-items: flex-start;
+    }
+
+    @media (min-width: 1280px) {
+      margin-top: 70px;
+    }
+  }
+
+  &__logo {
+    width: 38px;
+    height: 38px;
+  }
+
+  &__title {
+    margin: 50px 0 0;
+
+    font-weight: 500;
+    font-size: 24px;
+    line-height: 29px;
+    text-align: center;
+
+    @media (min-width: 568px) {
+      margin-top: 40px;
+    }
+  }
+
+  &__form {
+    width: 100%;
+
+    display: flex;
+    //align-items: flex-start;
+    flex-direction: column;
+  }
+
+  &__link {
+    margin: 14px auto 0;
+
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 15px;
+    color: #A0A0A0;
+    text-align: center;
+
+    &-span {
+      margin-left: 7px;
+      color: #000000;
+    }
+
+    @media (min-width: 568px) {
+      margin-top: 16px;
+
+      font-size: 14px;
+      line-height: 17px;
+    }
+  }
+}
