+++ src/components/Pages/Main/Main.scss
