+++ src/components/Pages/Movies/Movies.scss
