+++ src/components/Sections/BurgerNav/BurgerNav.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import {Link} from "react-router-dom";
+
+import './BurgerNav.scss';
+
+const BurgerNav = ({links, active, setActive}) => {
+    return (
+        <>
+            <div className={active ? 'burger-menu__blur' : 'burger-menu__blur burger-menu__blur_disable'}/>
+            <div className={active ? 'burger-menu' : 'burger-menu burger-menu_disable'}
+                 onClick={() => setActive(!active)}>
+
+                <div className='burger-menu__content' onClick={evt => evt.stopPropagation()}>
+                    <button className='burger-menu__button' onClick={() => setActive(false)}/>
+
+                    <ul className='burger-menu__list'>
+                        {links.map(link =>
+                            <li className='burger-menu__list_item'>
+                                <Link
+                                    className='burger-menu__link'
+                                    path={link.path}
+                                >
+                                    {link.title}
+                                </Link>
+                            </li>
+                        )}
+                    </ul>
+                </div>
+            </div>
+        </>
+    );
+};
+
+export default BurgerNav;
