+++ src/components/Sections/BurgerNav/BurgerNav.scss
@@ -1,0 +1,95 @@
+.burger-menu {
+  position: fixed;
+  right: 0;
+  top: 0;
+
+  width: 100vw;
+  height: 100vh;
+
+  z-index: 10;
+
+  transform: translate(0, 0);
+  transition: all 1.5s;
+
+  @media (min-width: 768px) {
+    max-width: 520px;
+  }
+
+  &_disable {
+    transform: translate(100vw, 0);
+    transition: all 1.5s;
+  }
+
+  &__blur {
+    position: fixed;
+    top: 0;
+    left: 0;
+
+    width: 100%;
+    height: 100vh;
+
+    background-color: rgba(0, 0, 0, 0.23);
+    z-index: 5;
+
+    transform: translate(0, 0);
+    transition: 1.5s ease-in-out;
+
+    &_disable {
+      transform: translate(-100vw, 0);
+      transition: 1.5s ease-in-out;
+    }
+  }
+
+  &__content {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: #ffffff;
+
+    transition: all 0.5s;
+  }
+
+  &__button {
+    position: absolute;
+    top: 13px;
+    right: 13px;
+
+    width: 18px;
+    height: 18px;
+
+    background-image: url('../../../images/close.svg');
+    background-position: center;
+    background-repeat: no-repeat;
+    background-color: transparent;
+
+    border: none;
+    cursor: pointer;
+  }
+
+  &__list {
+    margin: 144px 0 0;
+    padding: 0;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+
+    list-style-type: none;
+
+    &_item {
+      margin-bottom: 24px;
+    }
+  }
+
+  &__link {
+    margin: 0;
+    padding: 0;
+
+    font-weight: 500;
+    font-size: 18px;
+    line-height: 21px;
+
+    &:hover {
+      border-bottom: 2px #000000 solid;
+    }
+  }
+}
