+++ src/components/Sections/Header/Header.scss
@@ -1,0 +1,19 @@
+.header {
+  position: relative;
+  padding: 18px 14px;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+
+  @media (min-width: 568px) {
+    padding-left: 30px;
+    padding-right: 30px;
+  }
+
+  @media (min-width: 1050px) {
+    padding-left: 70px;
+    padding-right: 70px;
+  }
+}
