+++ src/components/Sections/MoviesCardList/MoviesCard/MoviesCard.scss
@@ -1,0 +1,72 @@
+.card {
+  max-width: 300px;
+
+  box-shadow: 0 -5px 5px rgba(4, 4, 4, 0.03);
+
+  @media (min-width: 750px) {
+    max-width: 339px;
+  }
+
+  @media (min-width: 1280px) {
+    max-width: 364px;
+  }
+
+  &__header {
+    padding: 14px;
+
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+  }
+
+  &__button-save {
+    width: 30px;
+    height: 30px;
+
+    border: none;
+    background-image: url('../../../../images/save-dis.svg');
+    background-position: center;
+    background-repeat: no-repeat;
+    background-color: transparent;
+
+    cursor: pointer;
+  }
+
+  &__button-save:active {
+    background-image: url('../../../../images/save-act.svg');
+  }
+
+  &__title {
+    margin: 0;
+    padding: 0;
+
+    font-weight: 500;
+    font-size: 15px;
+    line-height: 18px;
+    text-align: left;
+  }
+
+  &__subtitle {
+    font-weight: 400;
+    font-size: 11px;
+    line-height: 13px;
+    color: #A0A0A0;
+  }
+
+  &__img {
+    max-width: 100%;
+    height: 168px;
+
+    object-position: center;
+    object-fit: cover;
+
+    @media (min-width: 750px) {
+      height: 190px;
+    }
+
+    @media (min-width: 1280px) {
+      height: 203px;
+    }
+  }
+}
