+++ src/components/Sections/MoviesCardList/MoviesCardList.scss
