+++ src/components/Sections/Techs/Techs.js
