+++ src/components/Sections/Techs/Techs.scss
